Added Hitscan Nospread

+++ Fedoraware/Fedoraware-TF2/src/SDK/SDK.h
@@ -514,7 +514,7 @@ namespace Utils
 		fnRandomSeed(iSeed);
 	}
 
-	__inline float PlatFloatTime()
+	__inline double PlatFloatTime()
 	{
 		static auto fnPlatFloatTime = reinterpret_cast<float(*)()>(GetProcAddress(GetModuleHandleA(TIER0_DLL), "Plat_FloatTime"));
 		return fnPlatFloatTime();
